+++ cfdm/read_write/netcdf/netcdfread.py
@@ -11,6 +11,7 @@ from functools import reduce
 from math import log, nan, prod
 from numbers import Integral
 from os.path import isdir, isfile, join
+from pathlib import Path
 from typing import Any
 
 import numpy as np
@@ -9846,7 +9847,7 @@ class NetCDFRead(IORead, FieldChecker, NetCDFCheckerMixin):
 
         """
         # Strings (Paths)
-        if isinstance(dataset, str):
+        if isinstance(dataset, (str, Path)):
             return "path"
 
         # Check for a "virtual directory" (Mapper)